+++ server/app/services/ranked_service.go
@@ -5,7 +5,7 @@ import (
 	"osdtyp/app/entity"
 )
 
-func (s *ServiceLayer) JoinRankedLobby(ctx context.Context, userid uint64, duration entity.LobbyType) error {
+func (s *ServiceLayer) JoinRankedLobby(ctx context.Context, userid uint32, duration entity.LobbyType) error {
 	rank, err := s.db.GetRank(ctx, userid)
 	if err != nil {
 		return err
@@ -14,6 +14,6 @@ func (s *ServiceLayer) JoinRankedLobby(ctx context.Context, userid uint64, durat
 }
 
 // The game will be handled automatically using za core/
-func (s *ServiceLayer) GetRank(ctx context.Context, userid uint64) (uint16, error) {
+func (s *ServiceLayer) GetRank(ctx context.Context, userid uint32) (uint16, error) {
 	return s.db.GetRank(ctx, userid)
 }
